requirements.txt
asgiref==3.5.2
Django==3.2.14
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.1
